Add the dummy qmetal backend.
